+++ src/gfx/sprites.js
@@ -22,7 +22,7 @@ export {
   drawHeartburn, bubbleRadius, drawBubble, drawSpines, drawSpikeGuy,
   drawBeanBaron, drawBeanBomb, drawKurnuttaja, drawCroak,
   drawTorvi, drawTorahdys, drawPaarma, drawPisara, drawYokki, drawKarvapallo,
-  drawPaukkupoho, drawPyorre, drawKummitus,
+  drawPaukkupoho, drawPyorre, drawKummitus, drawKuura, drawKolikkovaras,
 } from './sprites/enemies.js';
 export { drawBoss, BOSS_SIZES, BOSS_PLANS, BOSS_LIMBS, FLOOR_REACH, bossSize } from './sprites/boss.js';
 export {